--- src/pages/wanyoung/Login/Login.scss
@@ -1,117 +1,0 @@
-.login {
-  .loginBox {
-    font-family: cursive;
-    border: 1px solid #dbdbdb;
-    flex-direction: column;
-    width: 35rem;
-    height: 40rem;
-    background-color: #fff;
-
-    .loginBoxForm {
-      height: 30rem;
-
-      .loginBoxLogo {
-        font-size: 5rem !important;
-      }
-
-      .loginInputWrapper {
-        height: 13rem;
-
-        .loginBoxInput {
-          background-color: #FAFAFA;
-          border: 1px solid rgb(219, 219, 219, 0.7);
-          width: 26rem;
-          height: 3.6rem;
-          border-radius: 0.3rem;
-          text-indent: 1rem;
-          font-size: 1.4rem !important;
-
-          &:focus {
-            outline: 0.8px solid rgba(219, 219, 219, 1.0);
-          }
-
-          &::placeholder {
-            color: #8e8e8e;
-            font-size: 1.2rem !important;
-          }
-        }
-
-        .loginBoxBtn {
-          cursor: pointer;
-          background-color: #0095f6;
-          width: 26rem;
-          height: 3.6rem;
-          outline: none;
-          border: none;
-          color: #fff;
-          font-weight: 600;
-          font-size: 1.3rem !important;
-          border-radius: 0.3rem;
-        }
-      }
-
-    }
-  }
-
-  .loginBoxLine {
-    .loginBoxHr {
-      width: 12rem;
-      background-color: #dbdbdb;
-      border: none;
-      height: 0.1rem;
-    }
-
-    .loginBoxOr {
-      color: #8e8e8e;
-      margin: 0 1rem;
-    }
-  }
-
-
-
-
-
-  .loginBoxInputBottom {
-    height: 8rem;
-
-    .bottomBtn {
-      width: inherit;
-
-      .facebookText {
-        font-size: 1.4rem !important;
-      }
-    }
-
-    .forgotPasswd {
-      font-size: 1.4rem !important;
-      color: #00376b;
-      text-decoration: none;
-    }
-  }
-
-  .loginBoxBottom {
-    background-color: #fff;
-    border: 1px solid rgb(219, 219, 219, 0.7);
-    width: 35rem;
-    height: 8rem;
-
-    .forgotPasswd {
-      font-size: 1.4rem !important;
-      color: #00376b;
-      text-decoration: none;
-    }
-
-    .loginBoxBottomText {
-      width: fit-content;
-      font-weight: 650;
-      color: #0095f6;
-      text-decoration: none;
-      font-size: 1.4rem !important;
-    }
-  }
-
-  .disabled {
-    background-color: #0095f6;
-    opacity: 0.5;
-  }
-}
