+++ src/pages/wanyoung/Main/Main.scss
@@ -26,267 +26,8 @@
     }
   }
 
-  .navBar {
-    z-index: 100;
-    background-color: #fff;
-    position: fixed;
-    top: 0;
-    width: 100vw;
-    height: 6rem;
-    border-bottom: 1px solid rgba(223, 223, 210, 1.0);
-
-    .navBarInner {
-      width: 70vw;
-
-      .logo {
-        font-family: cursive;
-        font-size: 3rem !important;
-      }
-
-      .navBarInput {
-        outline: none;
-        border: none;
-        background: #FAFAFA;
-        color: #8e8e8e;
-        font-size: 1.5rem !important;
-        font-weight: 500;
-        text-indent: 3rem;
-        border-radius: 0.8rem;
-        width: 25rem;
-        height: 3.6rem;
-        border: 1px solid rgba(223, 223, 210, 1.0);
-
-        &:focus {
-          border: 1px solid rgb(179, 179, 179);
-        }
-      }
-
-      .searchUserResult {
-        top: 5rem;
-        z-index: 200;
-        position: absolute;
-        width: 25rem;
-        height: 30rem;
-        background-color: #fff;
-        border-radius: 1rem;
-        max-height: 30rem;
-        overflow: auto;
-
-        .searchResult {
-          div {
-            p {
-              font-size: 1.5rem !important;
-            }
-          }
-
-          &:hover {
-            background-color: #d7d7d7;
-          }
-        }
-      }
-
-      .navBarItem {
-        width: 25rem;
-
-        .hiddenNavBar {
-          border-radius: 1rem;
-          left: 13rem;
-          top: 5rem;
-          background-color: #fff;
-          width: 20rem;
-          height: 20rem;
-          box-shadow: 1px 1px 1px 1px rgba(223, 223, 210, 1.0);
-
-          .hiddenNavItem {
-            height: 15rem;
-
-            .hiddenNavImgWrapper {
-              width: 19rem;
-              height: 3rem;
-
-              &:hover {
-                background-color: #d7d7d7;
-              }
-
-              .hiddenNavBarImg {
-                width: 2.4rem;
-                height: 2.4rem;
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  // .glass {
-  //   position: absolute;
-  //   top: 1rem;
-  //   left: 1.2rem;
-  //   font-size: 1.5rem;
-  //   color: #8e8e8e;
-  // }
-
   .mainContent {
     margin-top: 6rem;
     justify-content: center;
-
-    .leftContent {
-      .leftContentNewStory {
-        flex-wrap: nowrap;
-        border-radius: 0.8rem;
-        margin-top: 2rem;
-        width: 47rem;
-        min-width: 47rem;
-        overflow: auto;
-        height: 10rem;
-        background-color: #fff;
-        border: 1px solid rgba(223, 223, 210, 1.0);
-
-        .newStoryInnerBox {
-          width: 6rem;
-          height: 6rem;
-
-          span {
-            font-size: 1.3rem !important;
-          }
-
-          .newStoryUserImg {
-            top: 0rem;
-          }
-        }
-      }
-
-      .leftContentFeedBox {
-        margin-top: 2rem;
-        border-radius: 0.8rem;
-        width: fit-content;
-        background-color: #fff;
-        border: 1px solid rgba(223, 223, 210, 1.0);
-
-        .userText {
-          font-size: 1.5rem !important;
-          font-weight: 550;
-        }
-
-        .feedBoxTitle {
-          width: inherit;
-          height: 6rem;
-
-          .userTitleImg {
-            width: 2.9rem;
-            height: 2.9rem;
-            border-radius: 1.45rem;
-          }
-
-          .userText {
-            font-size: 1.5rem !important;
-            font-weight: 550;
-          }
-        }
-
-        .feedBoxImg {
-          width: 47rem;
-          height: 58rem;
-        }
-
-        .feedItems {
-          height: 6rem;
-
-          .feedInnerItems {
-            justify-content: flex-end;
-            align-items: center;
-            width: 100%;
-          }
-        }
-      }
-
-      .feedComment {
-        max-height: 7rem;
-        height: 7rem;
-        overflow: auto;
-
-        li {
-          font-size: 1.2rem !important;
-        }
-      }
-
-      .feedInput {
-        border-top: 1px solid rgba(223, 223, 210, 1.0);
-        width: 100%;
-        height: 4rem;
-
-        button {
-          cursor: pointer;
-          font-weight: 600;
-          outline: none;
-          border: none;
-          background: none;
-          width: 5rem;
-          color: #0095f6;
-        }
-
-        input {
-          font-size: 1.6rem !important;
-          border: none;
-          outline: none;
-          width: 100%;
-
-          &::placeholder {
-            font-size: 1.6rem !important;
-            color: #8e8e8e;
-          }
-        }
-      }
-    }
-
-    .rightContent {
-      margin-top: 4rem;
-      background-color: #FAFAFA;
-      width: 35rem;
-      height: 100px;
-
-      .rightContentWrapper {
-        width: 35rem;
-
-        .rightContentUserName {
-          font-size: 1.5rem !important;
-          font-weight: 600;
-        }
-
-        .rightContentUserRecommend {
-          color: #8e8e8e;
-          font-weight: 600;
-
-          p {
-            font-size: 1.4rem !important;
-          }
-        }
-
-        .recommendUserList {
-          padding: 3rem;
-
-          .recommendUserListLi {
-            width: 32rem;
-            height: 5rem;
-
-            .rightContentUserNameDesc {
-              font-size: 1.3rem !important;
-              color: #8e8e8e;
-            }
-
-            .recommendUserListFollow {
-              font-size: 1.5rem !important;
-              color: #0095f6;
-            }
-
-          }
-        }
-
-      }
-    }
-
-
   }
-
 }
