+++ src/pages/wanyoung/components/LeftContent.scss
@@ -1,0 +1,40 @@
+.leftContent {
+
+  .feedComment {
+    max-height: 7rem;
+    height: 7rem;
+    overflow: auto;
+
+    li {
+      font-size: 1.2rem !important;
+    }
+  }
+
+  .feedInput {
+    border-top: 1px solid rgba(223, 223, 210, 1.0);
+    width: 100%;
+    height: 4rem;
+
+    button {
+      cursor: pointer;
+      font-weight: 600;
+      outline: none;
+      border: none;
+      background: none;
+      width: 5rem;
+      color: #0095f6;
+    }
+
+    input {
+      font-size: 1.6rem !important;
+      border: none;
+      outline: none;
+      width: 100%;
+
+      &::placeholder {
+        font-size: 1.6rem !important;
+        color: #8e8e8e;
+      }
+    }
+  }
+}
