+++ src/pages/wanyoung/components/MainNewStory.scss
@@ -1,0 +1,11 @@
+.leftContentNewStory {
+  flex-wrap: nowrap;
+  border-radius: 0.8rem;
+  margin-top: 2rem;
+  width: 47rem;
+  min-width: 47rem;
+  overflow: auto;
+  height: 10rem;
+  background-color: #fff;
+  border: 1px solid rgba(223, 223, 210, 1.0);
+}
