+++ src/pages/wanyoung/components/NewStoryImg.scss
@@ -1,0 +1,12 @@
+.newStoryInnerBox {
+  width: 6rem;
+  height: 6rem;
+
+  span {
+    font-size: 1.3rem !important;
+  }
+
+  .newStoryUserImg {
+    top: 0rem;
+  }
+}
